Revert "Silence auto-value warning"
This change is no longer needed, following http://r.android.com/943770
in April 2019, which upgraded external/dagger2 from a 2015 version of
auto-common to version 0.10
(https://github.com/google/auto/releases/tag/auto-common-0.10). This
auto-common version does not output code using
javax.annotation.Generated if it is is not available, so the module
that provides (java.xml.ws.annotation) it can be safely removed.

This is beneficial since the module is deprecated in OpenJDK 9 and is
removed in later versions.

This reverts commit 87068a5236e63687231098ff446d9186b94ff00d.

Bug: 62050818
Bug: 131683177
Test: EXPERIMENTAL_JAVA_LANGUAGE_LEVEL_9=true make droid
Change-Id: I08f9b0696f9c550789d101affb653c3e892da0f2