Restore checkerframework annotations
I3e4794368c1369ccb32f24e55df030afe6a784d7 reverted upstream
commit d0fcced7ecc9108d70ae2aea0ae3db00aacf3924 to move from
checkerframework annotations back to jsr305 annotations.  An
upcoming turbine change uses the @Nullable in a location that
is not valid for the jsr305 annotations but is valid for
the checkerframework annotations.  We don't have a real copy of
the checkerframework annotations in the tree, but we do have a
copy packaged into dagger2-auto-value.  Restore the java/
directory to match upstream, and let dagger2-auto-value provide
the annotations for now.

Test: m turbine
Change-Id: If7f38e6b54aba2581c0e82a311a1d8e0d5722181